remove game python impl
