--- src/game/__init__.py
@@ -1,3 +1,0 @@
-from .game import Game, GameRS
-from .player import Player
-from .map import Map
