+++ src/sio/__init__.py
@@ -1,1 +1,2 @@
 from .job import JobManager, Job
+from .game import Game
